Added Calibrated Values as metrics

+++ algorithmFunction.h
@@ -1,7 +1,7 @@
 #include <stdbool.h>
 
 typedef struct DataEntry {
-    int temp;
+    float temp;
     float skinCond;
     float heartRate;
 } DataEntry;
@@ -9,10 +9,10 @@ typedef struct DataEntry {
 //The data entry and whether the user is in an episode or not.
 int deterministicAlgorithm(DataEntry entry, bool episodeStat, int episodeCount, DataEntry normalStats);  
 
-bool tempCheckNE(int temp, int normalTemp);
+bool tempCheckNE(float temp, float normalTemp);
 bool skinCondCheckNE(float skinCond, float normalSkinCond);
 bool heartRateCheckNE(float heartRate, float normalHeartRate);
 
-bool tempCheckE(int temp, int normalTemp);
+bool tempCheckE(float temp, float normalTemp);
 bool skinCondCheckE(float skinCond, float normalSkinCond);
 bool heartRateCheckE(float heartRate, float normalHeartRate);
